# improve worst-case performance of BTreeSet intersection v3