@KodrAus 2019-03-19
Hi @ssomers :wave:

Would you prefer to close #58577 and #59078 and look at this PR instead?

I haven't had a chance to look at what's changed since the changeset I approved, could you summarize the differences? It looks like this one does keep the `enum` for the strategy private which I think is an improvement on the v2 implementation.