@ssomers 2019-03-19
Sure. The difference with the approved state is:
- struct Intersection has **both fields replaced** by the enum, even though both implementations share the same iterator. I think it makes the code more readable and it simplifies changing/adding algorithms that don't need that field.
- The gut of both implementations is exposed (but hidden, again) as **unstable feature**, to allow the benchmark to compare them.
- Thus new benchmarks contrast the performance of both implementations with the actual implementation chosen by the size rule.
- The size rule favors the new implementation much more than before, but still less than looks optimal (on my system, for this benchmark, with these kind of elements). And if one of the sets is empty, the other set's iterator isn't even constructed.
- Sets are always ordered by size, because it simplifies code and because size_hint already had to choose the small set.
- The classic implementation no longer relies on Peekable (which greatly helps for union and difference, but not here).