@ssomers 2019-03-24
I configured Travis on [the separate repository](https://github.com/ssomers/rust_bench_btreeset_intersection) and the readings on the linux build on that (virtual) machine are quite similar (with much less stability, as one would expect). The ideal factor is < 1 higher accross the range of sizes.