@KodrAus 2019-03-25
Thanks for all your investigation work @ssomers!

Just for good measure, let's add another test case to `liballoc::tests::btree::set::test_intersection` that checks intersection of a very large set with a very small one so we can be sure it'll be using the other strategy. Something like:

```rust
check_intersection(&[11, 5000, 1, 3, 77, 8924, 103],
    &(0..1000).collect::<Vec<_>>(),
    &[11, 1, 3, 77, 103]);
```