@ssomers 2019-03-26
I thought the test cases [in the first PR](https://github.com/rust-lang/rust/pull/58577/files) were still there and they didn't even include non-subset samples. Rest assured, the proptest in the separate repository covered everything.

So are we there now? Nope, I just realized that is_subset/superset is really quite similar. I'll keep that out of this PR, but it might mean that the 16 becomes a named constant.