@ssomers 2019-03-27
I cooked up a similar change to the implementation of set difference. It only needs half of the peekables, and it benefits from the same performance boost as with intersection if the right hand set is huge. I cooked up similar code resulting in:

before:

    test btree::set::difference_random_100_vs_100            ... bench:         914 ns/iter (+/- 27)
    test btree::set::difference_random_100_vs_10k            ... bench:      45,641 ns/iter (+/- 605)
    test btree::set::difference_random_10k_vs_100            ... bench:      63,561 ns/iter (+/- 512)
    test btree::set::difference_random_10k_vs_10k            ... bench:     205,747 ns/iter (+/- 5,792)
    test btree::set::difference_staggered_100_vs_100         ... bench:         989 ns/iter (+/- 19)
    test btree::set::difference_staggered_100_vs_10k         ... bench:      35,919 ns/iter (+/- 418)
    test btree::set::difference_staggered_10k_vs_10k         ... bench:      93,716 ns/iter (+/- 1,358)

after:

    test btree::set::difference_random_100_vs_100            ... bench:         829 ns/iter (+/- 14)
    test btree::set::difference_random_100_vs_10k            ... bench:       2,871 ns/iter (+/- 112)
    test btree::set::difference_random_10k_vs_100            ... bench:      60,580 ns/iter (+/- 675)
    test btree::set::difference_random_10k_vs_10k            ... bench:     188,386 ns/iter (+/- 1,727)
    test btree::set::difference_staggered_100_vs_100         ... bench:         863 ns/iter (+/- 21)
    test btree::set::difference_staggered_100_vs_10k         ... bench:       2,673 ns/iter (+/- 70)
    test btree::set::difference_staggered_10k_vs_10k         ... bench:      83,875 ns/iter (+/- 2,977)

Do you want me to commit it here or later? (or not at all...)