@KodrAus 2019-03-27
This is looking great! Thanks for giving these methods some TLC @ssomers.

> Do you want me to commit it here or later? (or not at all...)

Yeh I think we can roll difference into this PR as well while we're working on these set operations. I'm happy with the implementation of intersection now.