@ssomers 2019-03-28
I meant "push" instead of "commit", but github lists the commits according to time committed locally anyway. Confusing...

Anyways, it's all here now, and nothing changed to the implementation of intersection.