@KodrAus 2019-03-29
Alrighty, this PR introduces some additional complexity to `BTreeSet`, but is much more efficient when performing set operations between a large and small set, which I think is a reasonable case. There's no change in the public API. So let's merge this one in!

@bors r+